+++ src/Symfony/Component/Messenger/Command/FailedMessagesRetryCommand.php
@@ -50,7 +50,7 @@ class FailedMessagesRetryCommand extends AbstractFailedMessagesCommand implement
         private EventDispatcherInterface $eventDispatcher,
         private ?LoggerInterface $logger = null,
         ?PhpSerializer $phpSerializer = null,
-        private ?array $signals = null
+        private ?array $signals = null,
     ) {
         parent::__construct($globalReceiverName, $failureTransports, $phpSerializer);
     }
@@ -132,7 +132,7 @@ EOF
 
     public function getSubscribedSignals(): array
     {
-        return $this->signals ?? (\extension_loaded('pcntl') ? [\SIGTERM, \SIGINT] : []);
+        return $this->signals ?? (\extension_loaded('pcntl') ? [\SIGTERM, \SIGINT, \SIGQUIT] : []);
     }
 
     public function handleSignal(int $signal, int|false $previousExitCode = 0): int|false